# src/main/java/com/discoveryit/assessment/model/Planet.java
package com.discoveryit.assessment.model;

import javax.persistence.Column;
import javax.persistence.Entity;
import javax.persistence.Id;
import javax.persistence.Table;
import javax.validation.constraints.NotBlank;
import java.util.Objects;

@Entity
@Table(name = "Planet")
public class Planet {

    @Id
    @Column(name = "routeId")
    @NotBlank
    private String id;
    @NotBlank
    private String name;

    public Planet(String id, String name){
        this.id = id;
        this.name = name;
    }

    public Planet(){
    }

    public String getId() {
        return id;
    }

    public void setId(String id) {
        this.id = id;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        Planet planet = (Planet) o;
        return Objects.equals(id, planet.id) &&
                Objects.equals(name, planet.name);
    }
}